+++ mptest/CHANGES.md
@@ -5,6 +5,9 @@ Change history for MP-Test
 Since last release
 ------------------
 
+#### 3/7/18
+  - Replace clock()/etime() with tic()/toc() for timing.
+
 #### 1/30/18
   - Added `abs()` to output of failed tests for consistency when
     comparing complex numbers (affects display only, test were correct).